dev_v1.0.0

+++ internal/repository/history/pg/init.go
@@ -2,11 +2,17 @@ package pg
 
 import (
 	"context"
+	"errors"
 	errorCommon "github.com/ChoTracker-C23-PS308/ChoTracker-CC/common/error"
 	"github.com/ChoTracker-C23-PS308/ChoTracker-CC/common/sqlc"
 	hModel "github.com/ChoTracker-C23-PS308/ChoTracker-CC/internal/model/history"
+	uModel "github.com/ChoTracker-C23-PS308/ChoTracker-CC/internal/model/user"
 	auRepo "github.com/ChoTracker-C23-PS308/ChoTracker-CC/internal/repository/auth"
 	"github.com/jackc/pgx/v4"
+)
+
+var (
+	ErrUser_UserNotAuthorized = errors.New("CREATE_HISTORY.USER_NOT_AUTHORIZED")
 )
 
 type pgHistoryRepository struct {
@@ -14,7 +20,11 @@ type pgHistoryRepository struct {
 	huRepo  auRepo.Repository
 }
 
-func (r pgHistoryRepository) CreateHistory(ctx context.Context, arg hModel.AddHistory, au hModel.AuthHistory) (string, error) {
+func (r pgHistoryRepository) CreateHistory(ctx context.Context, arg hModel.AddHistory, au uModel.AuthUser) (string, error) {
+	if !au.IsSame(arg.Uid) {
+		return "", ErrUser_UserNotAuthorized
+	}
+
 	id, err := r.querier.CreateHistory(ctx, sqlc.CreateHistoryParams(arg))
 	if err == pgx.ErrNoRows {
 		return "", errorCommon.NewNotFoundError("History not found")
@@ -22,7 +32,11 @@ func (r pgHistoryRepository) CreateHistory(ctx context.Context, arg hModel.AddHi
 	return id, err
 }
 
-func (r pgHistoryRepository) DeleteHistory(ctx context.Context, id string, au hModel.AuthHistory) error {
+func (r pgHistoryRepository) DeleteHistory(ctx context.Context, uid string, id string, au uModel.AuthUser) error {
+	if !au.IsSame(uid) {
+		return ErrUser_UserNotAuthorized
+	}
+
 	err := r.querier.DeleteHistory(ctx, id)
 	if err == pgx.ErrNoRows {
 		return errorCommon.NewNotFoundError("History not found")
@@ -30,7 +44,11 @@ func (r pgHistoryRepository) DeleteHistory(ctx context.Context, id string, au hM
 	return err
 }
 
-func (r pgHistoryRepository) GetHistory(ctx context.Context, uid string, au hModel.AuthHistory) ([]hModel.History, error) {
+func (r pgHistoryRepository) GetHistory(ctx context.Context, uid string, au uModel.AuthUser) ([]hModel.History, error) {
+	if !au.IsSame(uid) {
+		return []hModel.History{}, ErrUser_UserNotAuthorized
+	}
+
 	histories, err := r.querier.GetHistory(ctx, uid)
 	if err != nil {
 		return nil, err
